requirements.txt
pandas
requests
PyGithub
habanero
python-doi
arxiv
python-dotenv
backoff
paperscraper
backoff
supabase
httpx